Setup the server and the database
Built the server with initialization of database, token, and application configurations.

Established a functional and secured connection with the database.

Utilized .env file handling to keep sensitive and personal data.

Make use of requirements.txt file to list the required modules in running and developing the application and for easy libraries installation using pip.

Compiled the module instances in a single file for more contained codebase.