Generate, check, and revoke JWT token
Added application configurations for better and protected user experience.

Allowed token generation, checking, and revocation for enhanced security.

Created a logout authentication route for users to be able to change their account.

Removed "deploy" script in client package.json as it will not be deployed in GitHub pages and there is no deployment platform yet.

Redesigned code for enhanced readability and functionality.
• Changed class methods of AuthService class to instance methods.