Improve code structure
Removed unnecessary module imports for cleaner code.

Excluded identity from session storage to be submitted in the JSON payload of authentication request as it is not needed by the server.

Removed email field insertion in the database if it is empty to fix the bug where duplicate empty email fields recognized as  HTTP 409.

Remove email field in the database query of user registration to prevent looking for empty emails when it is not sent by the client.