Enable task modification
Allowed users to modify task details as well as set respond status to change necessary details and tell other users that the specific task is being responded to.

Added username to the created task to uniquely identify the creator of the task for entity relationship.

Added update user data function to handle database updates for different scenarios.

Modified user input filtration to exclude submitting keys that is null or not needed.