Enable task respond
Allowed users to respond to a task for all the clients to see which task is being responded to and which is not.

Added a toggle respond button that changes color and can be turned on and off to control the respond status of the task.

Added update task function in the client to modify the task in the server.

Fixed the rendering of tasks to prevent duplicating the task whenever the status changes.

Fixed task emitting error by converting the id of the updated task from ObjectId to string to be able to send to the client.