Enable task upvote
Allowed committing and counting of task upvotes to sort tasks from most important to least.

Changed boolean respondent to username respondent to identify who is in charge of responding to the task.

Changed upvotes count to list of upvotes to identify who upvoted the task.

Set contact and description default to None to prevent creating empty fields in database.

Added task date creation to identify how long the task has been requested.

Modified contact field to be not required as it may not always be needed to some scenarios.